server/gdb-server.c: make thread_id 32-bit value for ARC
For 32-bit ARC target->rtos->current_thread value is 32-bit and is
like 0x8xxxxxxx, but thread_id 64-bit and is like 0xffffffff8xxxxxxx.
This leads to "fake_step" case even if first 32 bits of thread_id is
equal to target->rtos->current_thread.

This commit fixes break/continue functionality for 32-bit systems.

+++ src/server/gdb_server.c
@@ -2760,7 +2760,7 @@ static bool gdb_handle_vcont_packet(struct connection *connection, const char *p
 		if (strncmp(parse, "s:", 2) == 0) {
 			struct target *ct = target;
 			int current_pc = 1;
-			int64_t thread_id;
+			uint32_t thread_id;
 			char *endp;
 
 			parse += 2;
